feat: sign traffic between RS and AS
