+++ packages/css/config/uma/default.json
@@ -5,11 +5,49 @@
   ],
   "@graph": [
     {
+      "@id": "urn:solid-server:default:ServerConfigurator",
+      "@type": "ParallelHandler",
+      "handlers": [
+        {
+          "@id": "urn:solid-server:default:StatusDependantServerConfigurator",
+          "@type": "StatusDependantServerConfigurator",
+          "dependants": [
+            { "@id": "urn:solid-server:default:Fetcher" }
+          ],
+          "statusMap": [
+            {
+              "StatusDependantServerConfigurator:_statusMap_key": "listening",
+              "StatusDependantServerConfigurator:_statusMap_value": true
+            },
+            {
+              "StatusDependantServerConfigurator:_statusMap_key": "close",
+              "StatusDependantServerConfigurator:_statusMap_value": false
+            },
+            {
+              "StatusDependantServerConfigurator:_statusMap_key": "error",
+              "StatusDependantServerConfigurator:_statusMap_value": false
+            }
+          ]
+        }
+      ]
+    },
+    {
       "comment": "Returns the UMA ticket in case of an unauthorized request.",
       "@id": "urn:solid-server:default:UmaClient",
-      "@type": "UmaClientImpl",
-      "UmaClientImpl:pat": "MYPAT",
-      "UmaClientImpl:_options_maxTokenAge": 600
+      "@type": "UmaClient",
+      "baseUrl": {
+        "@id": "urn:solid-server:default:variable:baseUrl"
+      },
+      "keyGen": {
+        "@id": "urn:solid-server:default:JwkGenerator"
+      },
+      "fetcher": {
+        "@id": "urn:solid-server:default:Fetcher",
+        "@type": "PausableFetcher",
+        "fetcher": {
+          "@type": "BaseFetcher"
+        }
+      }
     },
     {
       "@id": "urn:solid-server:default:OwnerUtil",
@@ -35,12 +73,15 @@
       "store": {
         "@id": "urn:solid-server:default:ResourceStore"
       },
-      "ownerUtil": {
-        "@id": "urn:solid-server:default:OwnerUtil"
-      },
       "umaIdStore": {
         "@id": "urn:solid-server:default:UmaIdStore",
         "@type": "MemoryMapStorage"
+      },
+      "ownerUtil": {
+        "@id": "urn:solid-server:default:OwnerUtil"
+      },
+      "umaClient": {
+        "@id": "urn:solid-server:default:UmaClient"
       }
     },
     {
