+++ packages/css/src/identity/configuration/JwksKeyHolder.ts
@@ -1,4 +1,3 @@
-
 import * as jose from 'jose';
 
 /**
