+++ packages/css/src/uma/ResourceRegistrar.ts
@@ -1,8 +1,8 @@
 import type { ResourceIdentifier, MonitoringStore, KeyValueStorage } from '@solid/community-server';
-import { AccessMode, AS, getLoggerFor, StaticHandler } from '@solid/community-server';
+import { AS, getLoggerFor, StaticHandler } from '@solid/community-server';
 import { OwnerUtil } from '../util/OwnerUtil';
-import { fetchUmaConfig } from './util/UmaConfigFetcher';
 import { ResourceDescription } from '@solidlab/uma';
+import type { UmaClient } from '../uma/UmaClient';
 
 export class ResourceRegistrar extends StaticHandler {
   protected readonly logger = getLoggerFor(this);
@@ -11,6 +11,7 @@ export class ResourceRegistrar extends StaticHandler {
     protected store: MonitoringStore,
     protected umaIdStore: KeyValueStorage<string, string>,
     protected ownerUtil: OwnerUtil,
+    protected umaClient: UmaClient,
   ) {
     super();
 
@@ -27,12 +28,10 @@ export class ResourceRegistrar extends StaticHandler {
 
   protected async createResource(resource: ResourceIdentifier, owner: string): Promise<void> {
     const issuer = await this.ownerUtil.findIssuer(owner);
-    const pat = await this.ownerUtil.retrievePat(owner);
 
     if (!issuer) throw new Error(`Could not find UMA AS for resource owner ${owner}`);
-    if (!pat) throw new Error(`Could not find PAT for resource owner ${owner}`);
 
-    const { resource_registration_endpoint: endpoint } = await fetchUmaConfig(issuer);
+    const { resource_registration_endpoint: endpoint } = await this.umaClient.fetchUmaConfig(issuer);
 
     const description: ResourceDescription = {
       resource_scopes: [
@@ -46,17 +45,18 @@ export class ResourceRegistrar extends StaticHandler {
 
     this.logger.info(`Creating resource registration for <${resource.path}> at <${endpoint}>`);
 
-    try {
-      const resp = await fetch(endpoint, {
-        method: 'POST',
-        headers: {
-          'Authorization': `Bearer ${pat}`,
-          'Content-Type': 'application/json',
-          'Accept': 'application/json',
-        },
-        body: JSON.stringify(description),
-      });
+    const request = {
+      url: endpoint,
+      method: 'POST',
+      headers: {
+        'Content-Type': 'application/json',
+        'Accept': 'application/json',
+      },
+      body: JSON.stringify(description),
+    };
 
+    // do not await - registration happens in background to cope with errors etc.
+    this.umaClient.signedFetch(endpoint, request).then(async resp => {
       if (resp.status !== 201) {
         throw new Error (`Resource registration request failed. ${await resp.text()}`);
       }
@@ -68,41 +68,42 @@ export class ResourceRegistrar extends StaticHandler {
       }
       
       this.umaIdStore.set(resource.path, umaId);
-    } catch (error) {
+    }).catch(error => {
       // TODO: Do something useful on error
       this.logger.warn(
         `Something went wrong during UMA resource registration to create ${resource.path}: ${(error as Error).message}`
       );
-    }
+    });
   }
 
   protected async deleteResource(resource: ResourceIdentifier, owner: string): Promise<void> {
     const issuer = await this.ownerUtil.findIssuer(owner);
-    const pat = await this.ownerUtil.retrievePat(owner);
 
     if (!issuer) throw new Error(`Could not find UMA AS for resource owner ${owner}`);
-    if (!pat) throw new Error(`Could not find PAT for resource owner ${owner}`);
     
-    const { resource_registration_endpoint: endpoint } = await fetchUmaConfig(issuer);
+    const { resource_registration_endpoint: endpoint } = await this.umaClient.fetchUmaConfig(issuer);
 
     this.logger.info(`Deleting resource registration for <${resource.path}> at <${endpoint}>`);
 
     const umaId = await this.umaIdStore.get(resource.path);
+    const url = `${endpoint}/${umaId}`;
 
-    try {
+    const request = {
+      url,
+      method: 'DELETE',
+      headers: {}
+    };
+
+    // do not await - registration happens in background to cope with errors etc.
+    this.umaClient.signedFetch(endpoint, request).then(async _resp => {
       if (!umaId) throw new Error('Trying to delete unknown/unregistered resource; no UMA id found.');
 
-      const resp = await fetch(`${endpoint}/${umaId}`, {
-        method: 'DELETE',
-        headers: {
-          'Authorization': `Bearer ${pat}`,
-        }
-      });
-    } catch (error) {
+      await this.umaClient.signedFetch(url, request);
+    }).catch(error => {
       // TODO: Do something useful on error
       this.logger.warn(
         `Something went wrong during UMA resource registration to delete ${resource.path}: ${(error as Error).message}`
       );
-    }
+    });
   }
 }
