+++ packages/css/src/util/fetch/StatusDependant.ts
@@ -1,0 +1,7 @@
+
+/**
+ * Any object of which a status can be changed.
+ */
+export interface StatusDependant<T> {
+  changeStatus(status: T): Promise<void>, 
+}
