+++ packages/uma/config/routes/resources.json
@@ -6,7 +6,6 @@
     {
       "@id": "urn:uma:default:ResourceRegistrationHandler",
       "@type": "ResourceRegistrationRequestHandler",
-      "baseUrl": "urn:uma:variables:baseUrl",
       "resourceStore": { "@id": "urn:uma:default:ResourceRegistrationStore" }
     },
     {
