+++ packages/uma/config/tokens/factory/default.json
@@ -9,7 +9,7 @@
       "params_expirationTime": "5m",
       "params_aud": "solid",
       "issuer": { "@id": "urn:uma:variables:baseUrl" },
-      "keyholder": { "@id": "urn:uma:default:DefaultKeyholder" }
+      "keyGen": { "@id": "urn:uma:default:JwkGenerator" }
     }
   ]
 }
