--- packages/uma/src/util/FetchFactory.ts
@@ -1,9 +1,0 @@
-import {fetch} from 'cross-fetch';
-
-/**
- * Factory class yielding an authenticated fetch
- * for accessing some resource
- */
-export abstract class FetchFactory {
-    abstract getAuthenticatedFetch(clientId: string): typeof fetch;
-}
